Forks & creates new branch to work on this article
Trying to figure out how to get access to this article in my editor, Atom. Here's the message I get when I edit directly:
`You’re making changes in a project you don’t have write access to. We’ve created a fork of this project for you to commit your proposed changes to. Submitting a change will write it to a new branch in your fork, so you can send a pull request.`
Hopefully, this will get me access.